Move calcTextDimensions method to common/utils folder
