+++ src/common/utils/calc-text-dimensions.ts
